move HelixFlowError & Result
