+++ helixflow-core/src/lib.rs
@@ -8,6 +8,8 @@
 
 use std::any::Any;
 
+use uuid::Uuid;
+
 pub mod task;
 
 /// Marker trait for our data items
@@ -18,3 +20,30 @@ where
 {
     fn as_any(&self) -> &dyn Any;
 }
+
+#[derive(Debug, thiserror::Error)]
+pub enum HelixFlowError {
+    // The #[from] anyhow::Error will convert anything that offers `into anyhow::Error`.
+    #[error("backend error: {0}")]
+    BackendError(#[from] anyhow::Error),
+
+    #[error("created item does not match expectations: expected {expected:?}, got {actual:?}")]
+    Mismatch {
+        expected: Box<dyn HelixFlowItem>,
+        actual: Box<dyn HelixFlowItem>,
+    },
+
+    #[error("task id ({id:?}) is not a valid UUID v7")]
+    InvalidID { id: String },
+
+    #[error("404 No {itemtype} found with id {id}")]
+    NotFound { itemtype: String, id: Uuid },
+
+    #[error("Relationship between {left:?} and {right:?} contains Errors")]
+    RelationshipBetweenErrors {
+        left: Box<HelixFlowResult<Box<dyn HelixFlowItem>>>,
+        right: Box<HelixFlowResult<Box<dyn HelixFlowItem>>>,
+    },
+}
+
+pub type HelixFlowResult<T> = std::result::Result<T, HelixFlowError>;
